refactor(aggregator): move builders function to `enablers` submodule
